requirements.txt
xarray
pytest
netCDF4
cfgrib
datetime
numpy
pyyaml
cdsapi
pandas
scipy
argparse
asciitree
isodate
